feat(lambda-promtail): allow setting Lambda reserved concurrency
For the uninitiated, from https://docs.aws.amazon.com/lambda/latest/dg/configuration-concurrency.html

> Reserved concurrency – This represents the maximum number of
> concurrent instances allocated to your function. When a function has
> reserved concurrency, no other function can use that concurrency.
> Reserved concurrency is useful for ensuring that your most critical
> functions always have enough concurrency to handle incoming requests.
> Configuring reserved concurrency for a function incurs no additional
> charges.

Given this, being able to configure reserved concurrency will help in
AWS accounts that utilize Lambda heavily (not just for lambda-promtail)
in a few ways. Since reserved concurrency is the max concurrency a
function can run, it means that a deluge of events being sent to
lambda-promtail won't exhaust all of the Lambda concurrency for the
entire AWS account's region. On the other hand, since setting reserved
concurrency means that function will always have that amount of
concurrency reserved just for it, it can mitigate issues with delayed
log messages during periods of heavily Lambda usage since the
lambda-promtail function will be guaranteed to be able to handle N
number of events (N being the reserved concurrency) number of events at
any given time.

The default behavior in the Terraform AWS provider is to _not_ set any
reserved concurrency. It uses the `-1` value to denote this. To preserve
that behavior, the default value for the new input variable is `-1`.

For the keen-eyed, _provisioned_ concurrency is a different thing to
_reserved_ concurrency. _Reserved_ concurrency is configured on the
`aws_lambda_function` resource, but _provisioned_ concurrency is
configured via a separate resource
`aws_lambda_provisioned_concurrency_config`.